backend/requirements.txt
# Core Data Processing (compatible versions)
pandas==2.0.3
numpy==1.23.5

# Visualization
matplotlib==3.7.0
seaborn==0.12.0

# API Integration
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# NLP Libraries
scikit-learn==1.3.0
sentence-transformers==2.2.2

# LLM Integration
anthropic==0.8.0

# Utilities
tqdm==4.66.0

fastapi==0.110.0
uvicorn[standard]==0.27.1